Added validation, form, and input for weather API
I added a basic form input so that users can
search for different locations with the weather
API, and the API correctly gets the data for
those locations and also processes the .json
data to only contain the weather data we want.
Also added some stylings and the basic
webpage look. I plan on dynamically rendering
the webpage based on the processed weather data
and will re-render the webpage to contain icons
and visual representations of the weather from the
given location from the webAPI. I will also
check error handling to see if the location is
invalid/the API didn't return anything because
the location doesn't exist.