learned that to use async data must stay async
I had issues with my weatherRender module using
the data from the apiController module and that
was because the functions I made weren't async
so it didn't wait for the data from the
apiController, it just ran things synchronously
without waiting so thats why it kept showing
undefined because the data wasn't fetched yet.
When using async data, it is required to stay
in the asynchronous style because its never a
guarentee that the data will come in a
designated time or at all.