Added date-fns and added daily forecast dates
Had to convert the original date function to
ISO and THEN format it with date-fns first to
account for local timezones for accurate time/date.
Added styles to the day-date and changed the
border-radius to round out at the top of the
boxes. Need to add today's date to the
current weather box.